validation to userinputform
